feat: restrict to update/delete members and projects not created by user

+++ src/controller/project.ts
@@ -8,20 +8,12 @@ import { ProjectRepository } from '../repository/projectRepository';
 import { wrapError, wrapListResult, wrapResult } from '../utils/resWrapper';
 import { MemberRepository } from '../repository/memberRepository';
 
-const getUserProjectByReq = async (req: Request): Promise<Project | null> => {
-    const id = req.params.id as string;
-    const userId = getUserIdByReq(req);
-
-    return ProjectRepository.getProjectByIdAndUserId(id, userId);
-};
-
 export const getProjects = async (req: Request, res: Response) => {
-        const userId = getUserIdByReq(req);
-        const projects = await ProjectRepository.getProjectsByUserId(userId);
+    const userId = getUserIdByReq(req);
+    const projects = await ProjectRepository.getProjectsByUserId(userId);
 
-        return res.json(wrapListResult<Project>(projects));
-    }
-;
+    return res.json(wrapListResult<Project>(projects));
+};
 
 export const createProjectValidation = [
     body('name').trim().notEmpty(),
@@ -41,10 +33,20 @@ export const updateProjectValidation = [
     validateResult,
 ];
 export const updateProject = async (req: Request, res: Response) => {
+    const userId = getUserIdByReq(req);
     const { name }: { name: string } = req.body;
-    const project = await getUserProjectByReq(req);
+    const id = req.params.id as string;
+
+    const project = await ProjectRepository.getProjectByIdAndUserId(id, userId);
     if (!project) {
         return res.status(StatusCode.ClientErrorNotFound).json(wrapError('Project is not found'));
+    }
+
+    console.log(userId, project);
+
+    if (project.ownerId !== userId) {
+        console.log(111);
+        return res.status(StatusCode.ClientErrorUnauthorized).json(wrapError('Ask the project owner to update this project'));
     }
 
     const updatedProject = await ProjectRepository.updateProject(project.id, name);
@@ -53,9 +55,16 @@ export const updateProject = async (req: Request, res: Response) => {
 };
 
 export const deleteProject = async (req: Request, res: Response) => {
-    const project = await getUserProjectByReq(req);
+    const userId = getUserIdByReq(req);
+    const id = req.params.id as string;
+
+    const project = await ProjectRepository.getProjectByIdAndUserId(id, userId);
     if (!project) {
         return res.status(StatusCode.ClientErrorNotFound).json(wrapError('Project is not found'));
+    }
+
+    if (project.ownerId !== userId) {
+        return res.status(StatusCode.ClientErrorUnauthorized).json(wrapError('Ask the project owner to delete this project'));
     }
 
     const deletedProject = await ProjectRepository.deleteProject(project.id);
